requirements.txt
python-telegram-bot
selenium
schedule
python-dotenv
requests
